refactor(api): move test in prescription/campaign/acceptance/application/campaign-detail-route_test.js
