--- api/tests/acceptance/application/campaigns/campaign-management-controller_test.js
@@ -1,32 +1,0 @@
-import { createServer } from '../../../../server.js';
-import { expect } from '../../../test-helper.js';
-import { databaseBuilder } from '../../../tooling/databases.js';
-import { generateAuthenticatedUserRequestHeaders } from '../../../tooling/test-utils/http-server.js';
-
-describe('Acceptance | API | Campaign Management Controller', function () {
-  let server;
-
-  beforeEach(async function () {
-    server = await createServer();
-  });
-
-  describe('GET /api/admin/campaigns/{id}', function () {
-    it('should return the campaign details', async function () {
-      // given
-      const campaign = databaseBuilder.factory.buildCampaign();
-      const user = databaseBuilder.factory.buildUser.withRole();
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject({
-        method: 'GET',
-        url: `/api/admin/campaigns/${campaign.id}`,
-        headers: generateAuthenticatedUserRequestHeaders({ userId: user.id }),
-      });
-
-      // then
-      expect(response.statusCode).to.equal(200);
-      expect(response.result.data.id).to.equal(campaign.id.toString());
-    });
-  });
-});
