+++ api/tests/prescription/campaign/acceptance/application/campaign-detail-route_test.js
@@ -530,4 +530,24 @@ describe('Acceptance | API | campaign-detail-route', function () {
       expect(response.result.data).to.have.lengthOf(2);
     });
   });
+
+  describe('GET /api/admin/campaigns/{id}', function () {
+    it('should return the campaign details', async function () {
+      // given
+      const campaign = databaseBuilder.factory.buildCampaign();
+      const user = databaseBuilder.factory.buildUser.withRole();
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'GET',
+        url: `/api/admin/campaigns/${campaign.id}`,
+        headers: generateAuthenticatedUserRequestHeaders({ userId: user.id }),
+      });
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.result.data.id).to.equal(campaign.id.toString());
+    });
+  });
 });
